Notifications, Object Handling
Added a notification window to display when there is an incorrect object is placed in the system, and to display the message when the drawer needs to be emptied. The RFID functionality was also expanded, so that once the system checks all the slots and detects the objects, it determines flags for correct/incorrect color/shape and sends that information to the notification window to handle the errors. The game logic is fully implemented, but not fully tested. 

TODO: 
test current code and refine
implement audio control using GPIO into the Notif window error handling (receiving from game logic)
volume GUI control on settings and game pages
hardcode RFID tag IDs